+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2023-06-24
+### Added
+- Reverse Attachment Order v1.0.0
+- Recursive Hierarchy Labels v1.3.1
+- Changelog file: Get Table IDs
+
+### Changed
+- Get Table IDs v1.0.4 released, with support for comment block around field properties
+
 ## 2023-02-25
 ### Added
 - Get Table IDs